Tied add.html to backend to provide data to server
Provides data to fit the POJO abstraction to hold the added user data, but there are still some things that are missing.
POJO requires translation of certain data to be able to be saved, so that needs to be fixed.
Example is daysOfWeek, where it requires inserting a boolean array rather than being able to accept raw data from the frontend.